page rearrangement and framework implementation
