+++ src/app/(intro)/intro/layout.tsx
@@ -1,0 +1,11 @@
+export default function IntroLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="fr">
+      <body className="bg-black text-white">{children}</body>
+    </html>
+  );
+}
