+++ src/app/(site)/competences/page.tsx
