+++ src/app/globals.css
@@ -20,7 +20,16 @@
 }
 
 body {
-  background: var(--background);
+  /* background: var(--background); */
   color: var(--foreground);
   font-family: Arial, Helvetica, sans-serif;
+}
+
+.border-frame {
+  border-width: 34px;
+  border-image-source: url("/texture.svg");
+  border-image-slice: 30;
+  border-image-repeat: stretch;
+  border-style: solid;
+  
 }
